Convert provider_networks attrib to array
With node['quantum']['ovs']['provider_networks'] defaulting
to a Hash, it is impossible to remove the configuration for
ph-eth1. It can only be overridden.

This commit changes the syntax for defining the provider
networks a bit; they are now defined as flat Hashes in an
Array of provider networks. Ex:

[
  {
    "label" => "ph-eth1",
    "bridge" => "br-eth1",
    "vlans" => "1:1000"
  },
  {
    "label" => "ph-eth0",
    "bridge" => "br-eth0",
    "vlans" => "1001:1024"
  }
]