Fix for Issue 941
Process is to stamp the DB and then make a revision note indicating this is an RCBOPS deployment which is then used to ensure that the DB is NOT stamped on subsequent chef runs.

This commit adds new node attr hash to the ["neutron"]["db"]["stamp"]

The DB stamp will only be attempted when the RCBOPS deployment $REVISION
is not found in the neutron db history.

This also fixes some syntax issues we had in addition to making the recipe more legible.

Issue:
rcbops/chef-cookbooks#941

(cherry picked from commit ddfc0e05eb6f144e26789a3c0ccb111fe9f1b478)